# DRIVERS/server/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: sergejs.dyldin@yandex.ru
    name: Yandex Go Project Team
    url: https://github.com/GameXost/YandexGo_proj
  description: API for managing drivers and ride operations
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: DRIVER Service API
  version: "2.0"
servers:
- url: http://localhost:9096/
- url: https://localhost:9096/
security:
- BearerAuth: []
...
paths:
...
  /driver/profile:
    get:
      operationId: getDriverProfile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/driver_serviceDriver'
          description: Successfully retrieved driver profile
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rpcStatus'
          description: An unexpected error response.
      security:
      - BearerAuth: []
      summary: Get driver profile
      tags:
      - driver
    put:
      operationId: updateDriverProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/driver_serviceUpdateDriverProfileRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/driver_serviceDriver'
          description: Successfully updated driver profile
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rpcStatus'
          description: An unexpected error response.
      security:
      - BearerAuth: []
      summary: Update driver profile
      tags:
      - driver
      x-codegen-request-body-name: body
...
components:
  schemas:
    driver_serviceDriver:
      example:
        carNumber: carNumber
        carColor: carColor
        phone: phone
        carMark: carMark
        id: id
        email: email
        username: username
        carModel: carModel
      properties:
        id:
          description: Unic driver identificator
          type: string
        username:
          description: Driver's first_name
          type: string
        phone:
          description: contact number
          type: string
        carModel:
          description: Car model
          type: string
        email:
          description: Email address
          type: string
        carColor:
          description: Car color
          type: string
        carMark:
          description: Car manufacturer
          type: string
        carNumber:
          description: License plate number
          type: string
      type: object
...
    driver_serviceUpdateDriverProfileRequest:
      properties:
        id:
          description: Driver ID
          type: string
        driver:
          $ref: '#/components/schemas/driver_serviceDriver'
      type: object
...
    protobufAny:
      additionalProperties:
        type: object
      example:
        '@type': '@type'
      properties:
        '@type':
          type: string
      type: object
    rpcStatus:
      example:
        code: 0
        details:
        - '@type': '@type'
        - '@type': '@type'
        message: message
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
        details:
          items:
            $ref: '#/components/schemas/protobufAny'
          type: array
      type: object
  securitySchemes:
    BearerAuth:
      description: "JWT token as: Bearer <token>"
      in: header
      name: Authorization
      type: apiKey
...